fix: do not show duplicate tags on editing
when an entry was supposed to be added from a direct link and gets cancelled, then clicking on another entry for editing do not show the tags from search twice. Further, before fixation when the duplicated tag was removed both tags were get removed. Now it is working as it is supposing to.